make the device token poll interval configurable

+++ rauthy-models/src/response.rs
@@ -140,7 +140,7 @@ pub struct DeviceCodeResponse<'a> {
     pub verification_uri_complete: Option<String>,
     pub expires_in: u16,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub interval: Option<u32>,
+    pub interval: Option<u8>,
 }
 
 #[derive(Debug, PartialEq, Serialize, Deserialize, ToSchema)]
